split showdown and serebii script test files
no changes just moving
